changed package name from  model to entity
