+++ customerservice/src/main/java/com/project/customerservice/client/BankServiceFeignClient.java
@@ -1,12 +1,10 @@
 package com.project.customerservice.client;
 
-import com.project.customerservice.model.Bank;
+import com.project.customerservice.entity.Bank;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-
-import java.util.List;
 
 @FeignClient(name="BANK-MICROSERVICE", url = "http://localhost:8080")
 public interface BankServiceFeignClient {
