+++ customerservice/src/main/java/com/project/customerservice/entity/Customer.java
@@ -1,0 +1,42 @@
+package com.project.customerservice.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.validation.constraints.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class Customer {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @NotNull(message = "name is mandatory")
+    @Size(min = 3, message = "name should be minimum 3 characters")
+    private String name;
+
+    @NotNull(message = "Email is mandatory")
+    @Email(message = "Email should be valid")
+    private String email;
+
+    @NotNull(message = "Phone number is mandatory")
+    @Pattern(regexp = "^[0-9]{10}$", message = "Phone number must be exactly 10 digits")
+    private String phone;
+
+    @NotNull(message = "Address is mandatory")
+    @NotBlank(message = "Address cannot be blank")
+    private String address;
+
+    @NotNull(message = "Bank Id is mandatory")
+    private Long bankId;
+
+    private String accountNumber;
+}
