--- customerservice/src/main/java/com/project/customerservice/model/Bank.java
@@ -1,31 +1,0 @@
-package com.project.customerservice.model;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Bank {
-    private Long id;
-    private String name;
-    private String branch;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getBranch() {
-        return branch;
-    }
-
-    public void setBranch(String branch) {
-        this.branch = branch;
-    }
-}
