+++ customerservice/src/main/java/com/project/customerservice/repository/CustomerRepository.java
@@ -1,6 +1,6 @@
 package com.project.customerservice.repository;
 
-import com.project.customerservice.model.Customer;
+import com.project.customerservice.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
