+++ customerservice/src/main/java/com/project/customerservice/service/CustomerServiceImpl.java
@@ -4,10 +4,9 @@ import com.project.customerservice.Exception.ClientException;
 import com.project.customerservice.Exception.ResourceNotFoundException;
 import com.project.customerservice.Exception.ResoureExistsException;
 import com.project.customerservice.client.BankServiceFeignClient;
-import com.project.customerservice.model.Bank;
-import com.project.customerservice.model.Customer;
+import com.project.customerservice.entity.Bank;
+import com.project.customerservice.entity.Customer;
 import com.project.customerservice.repository.CustomerRepository;
-import feign.FeignException;
 import io.github.resilience4j.retry.annotation.Retry;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
