+++ loanservice/src/main/java/com/project/loanservice/entity/Bank.java
@@ -1,0 +1,15 @@
+package com.project.loanservice.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Bank {
+
+    private Long id;
+    private String name;
+    private String branch;
+}
