+++ loanservice/src/main/java/com/project/loanservice/entity/Customer.java
@@ -1,0 +1,19 @@
+package com.project.loanservice.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Customer {
+
+    private Long id;
+    private String name;
+    private String email;
+    private String phone;
+    private String address;
+    private Long bankId;
+    private String accountNumber;
+}
