--- loanservice/src/main/java/com/project/loanservice/model/Bank.java
@@ -1,15 +1,0 @@
-package com.project.loanservice.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Bank {
-
-    private Long id;
-    private String name;
-    private String branch;
-}
