--- loanservice/src/main/java/com/project/loanservice/model/Customer.java
@@ -1,20 +1,0 @@
-package com.project.loanservice.model;
-
-import jakarta.validation.constraints.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Customer {
-
-    private Long id;
-    private String name;
-    private String email;
-    private String phone;
-    private String address;
-    private Long bankId;
-    private String accountNumber;
-}
