+++ loanservice/src/main/java/com/project/loanservice/service/LoanService.java
@@ -1,6 +1,6 @@
 package com.project.loanservice.service;
 
-import com.project.loanservice.model.Loan;
+import com.project.loanservice.entity.Loan;
 
 public interface LoanService {
     Loan addLoan(Loan loan);
